PNPPK/UI/requirements.txt
matplotlib
PyQt5
PyQt5_sip
pyserial
PyYAML
